For loops
Changed the while loop into a for loop.

Added var links which takes the tag "a" from the html document. The for loop uses the local var i and initiates it at i=1. It will loop as long as i is less than or equal to 6, printing a string and the i that the loop is currently working on. Once the loop breaks, the script prints a string to the document. Since this is a practice for loop and I will most likely be erasing it shortly, I didn't add links to Lessons.html, so it won't run.