Breaks
Told ya I'd be erasing the previous code. I should ditch everything and become a nomad fortuneteller.

So, break and continue. This function will loop from 0, omit console.log for 3 and 5 (cuz continue) but continue (kek) the loop afterwards, and break will stop the loop at 7, even though the condition stated in for is i<10.